core: Store events in a calendar scheduler bucket in reverse chronological order
This improvement was actually introduced in ns v2.33, but lost in ns-3:
http://www.isi.edu/nsnam/ns/doc/node35.html

When a new event is inserted into a bucket, which is implemented as a
sorted list, it is more likely that new event has timestamp greater than
timestamps of previously inserted events.

Therefore, it is better to sort list in *reverse* chronological order.

This change makes CalendarScheduler outperform MapScheduler in benchmark:
./waf --run 'bench-simulator --cal=true --map=false'

For this reason, note about CalendarScheduler being slow is removed.